third_party/WebKit/LayoutTests/http/tests/devtools/compiler-source-mapping-debug-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests installing compiler source map in scripts panel.


Running: testSetBreakpoint
Script source was shown.
Script execution paused.
Call stack:
    0) handleClick (add-elements.js:15)
    1) clickButton (compiler-source-mapping-debug.js:18)
Script execution resumed.
